archi clean
creation of server_manager, a folder that contain all manager for better code separation
