+++ src/backend/server_manager/mod.rs
@@ -1,4 +1,5 @@
 pub mod account_manager;
 
+pub mod global_manager;
 pub mod pw_manager;
 pub mod vault_manager;
